requirements.txt
azure-core==1.23.1
azure-storage-blob==12.11.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
et-xmlfile==1.1.0
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
pandas==1.4.2
pycparser==2.21
pyodbc==4.0.32
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.35
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
WTForms==2.3.3
XlsxWriter==3.0.3
zipp==3.8.0



